Fix 36 bugs + watch mode + UI improvements (0.5.3)

+++ Source/Parsek/GhostSoftCapManager.cs
@@ -32,7 +32,7 @@ namespace Parsek
         /// Determines priority for a ghost recording. Lower priority = despawned first.
         /// Pure static for testability.
         /// </summary>
-        internal static GhostPriority ClassifyPriority(IPlaybackTrajectory traj, int loopCycleIndex)
+        internal static GhostPriority ClassifyPriority(IPlaybackTrajectory traj, long loopCycleIndex)
         {
             if (traj.LoopPlayback)
             {
